+++ src/kiosvn/kiosvn.h
@@ -51,7 +51,7 @@ public:
 
 private:
     KioSvnData*m_pData;
-    bool createUDSEntry( const QString& filename, const QString& user, long int size, bool isdir, time_t mtime, KIO::UDSEntry& entry);
+    bool createUDSEntry( const QString& filename, const QString& user, long long int size, bool isdir, time_t mtime, KIO::UDSEntry& entry);
     static QString makeSvnUrl(const KURL&url);
     bool getLogMsg(QString&);
 };
